Fix two of the three hacks currently in the code

+++ android/app/src/main/java/edu/berkeley/eecs/cfc_tracker/usercache/UserCache.java
@@ -7,10 +7,10 @@ public interface UserCache {
 
     class TimeQuery {
         int keyRes;
-        long startTs;
-        long endTs;
+        float startTs;
+        float endTs;
 
-        public TimeQuery(int keyRes, long startTs, long endTs) {
+        public TimeQuery(int keyRes, float startTs, float endTs) {
             this.keyRes = keyRes;
             this.startTs = startTs;
             this.endTs = endTs;
